% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_trinucleotide_composition.R
\name{calculate_trinucleotide_composition}
\alias{calculate_trinucleotide_composition}
\title{Calculate Trinucleotide Composition in a Genomic Region}
\usage{
calculate_trinucleotide_composition(grObject)
}
\arguments{
\item{grObject}{A GenomicRanges object representing the genomic region for which trinucleotide frequency will be calculated.}
}
\value{
A data frame containing the trinucleotide contexts and their corresponding counts.
}
\description{
This function computes the trinucleotide composition for a given genomic range object.
Both sense and antisense counts are considered.
}
\examples{
\dontrun{
gr <- GRanges("chr1", IRanges(1000, 2000))
result <- calculate_trinucleotide_composition(gr)
}
}
