% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gene_ids.R
\name{check_gene_aliases}
\alias{check_gene_aliases}
\title{Check for Gene Aliases}
\usage{
check_gene_aliases(hgnc, approved = TRUE)
}
\arguments{
\item{hgnc}{Character string. The HGNC symbol for which aliases are to be checked.}

\item{approved}{Logical. If TRUE, returns only the approved aliases; if FALSE, includes previous and alias symbols.}
}
\value{
A vector of unique gene aliases.
}
\description{
This function checks for aliases of a given gene symbol (HGNC) in the \code{mutmatch_geneAliases} data frame.
The function can return either approved aliases only or include previous and alias symbols.
}
