% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_regression_data_integrity.R
\name{check_regression_data_integrity}
\alias{check_regression_data_integrity}
\title{Check Regression Data Integrity}
\usage{
check_regression_data_integrity(
  mutation_table,
  formula =
    "MutationNumber ~ isTarget + CNA + isTarget:CNA + Mutation + offset(log(ntAtRisk))"
)
}
\arguments{
\item{mutation_table}{A data.table containing the dataset to check.}

\item{formula}{A character string representing the formula used in the regression model.
Default is "MutationNumber ~ isTarget + CNA + isTarget:CNA + Mutation + offset(log(ntAtRisk))".}
}
\value{
This function does not return a value. It will stop and throw an error message
if the data integrity checks fail.
}
\description{
This function checks the integrity of the data to be used in a regression model. It ensures that
the \code{isTarget} variable has two levels and checks the presence of all combinations of \code{isTarget}
and an interaction variable from the formula in the dataset. If any combination is missing, it
will throw an error.
}
\examples{
\dontrun{
check_regression_data_integrity(mutation_table)
}
}
