% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_dna_length.R
\name{convert_dna_length}
\alias{convert_dna_length}
\title{Convert Human-Readable DNA Length to Number of Nucleotides}
\usage{
convert_dna_length(dna_length)
}
\arguments{
\item{dna_length}{The human-readable DNA length (e.g., "0.5 Mb") or a numeric value.}
}
\value{
The DNA length in number of nucleotides.
}
\description{
This function takes a human-readable DNA length and converts it to the actual
number of nucleotides. Supported units are 'Mb', 'kb', and 'bp'.
If no unit is provided, it assumes the unit is 'bp'.
}
\examples{
\dontrun{
convert_dna_length("0.5 Mb") # Returns 500000
convert_dna_length("1000") # Returns 1000
convert_dna_length(1000) # Returns 1000
convert_dna_length("2 kb") # Returns 2000
}
}
