% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_process_pvalues.R
\name{correct_pvalues}
\alias{correct_pvalues}
\title{Correct P-Values to Remove Deflation or Inflation}
\usage{
correct_pvalues(real_pvalues, simulated_pvalues)
}
\arguments{
\item{real_pvalues}{A numeric vector of real p-values to be adjusted.}

\item{simulated_pvalues}{A numeric vector of simulated p-values used to generate the ECDF.}
}
\value{
A numeric vector of adjusted p-values.
}
\description{
This function corrects for deflation or inflation in real p-values by mapping them to their
corresponding quantiles within a distribution of simulated p-values. This adjustment is based on the empirical
cumulative distribution function (ECDF) derived from the simulated p-values, enhancing the statistical reliability
of the p-value estimates.
}
\examples{
\dontrun{
real_pvalues <- c(0.01, 0.05, 0.2)
simulated_pvalues <- runif(1000, 0, 1)
corrected_pvalues <- correct_pvalues(real_pvalues, simulated_pvalues)
}
}
