% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debias_selection_estimates.R
\name{debias_selection_estimates}
\alias{debias_selection_estimates}
\title{Debias Selection Estimates}
\usage{
debias_selection_estimates(estimates_original, estimates_simulated)
}
\arguments{
\item{estimates_original}{A data.table containing the original coefficient estimates.}

\item{estimates_simulated}{A data.table containing the simulated coefficient estimates.}
}
\value{
A data.table with corrected coefficient estimates for the isTarget variable and associated terms.
}
\description{
This function debiases selection estimates by taking the original estimates
and comparing them with simulated estimates. It corrects the coefficients
based on the median value of the simulated coefficients.
}
\examples{
\dontrun{
estimates_original <- data.table::data.table(
  coefName = c("isTarget", "CNA"),
  coef = c(0.1, 0.2)
)
estimates_simulated <- data.table::data.table(coefName = rep(c("isTarget", "CNA"),
  each = 4
), coef = runif(8))
debias_selection_estimates(estimates_original, estimates_simulated)
}
}
