% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_glm.R
\name{.bayesglm_fit}
\alias{.bayesglm_fit}
\title{A Wrapper Function for arm::bayesglm.fit}
\usage{
.bayesglm_fit(
  x,
  y,
  weights = rep(1, NROW(y)),
  family = gaussian(),
  offset = rep(0, NROW(y)),
  start = NULL,
  mustart = NULL,
  etastart = NULL,
  control = list(maxit = 100, epsilon = 1e-07)
)
}
\arguments{
\item{x, y}{For \code{glm}:
    logical values indicating whether the response vector and model
    matrix used in the fitting process should be returned as components
    of the returned value.

    For \code{glm.fit}: \code{x} is a design matrix of dimension \code{n
      * p}, and \code{y} is a vector of observations of length \code{n}.
  }

\item{weights}{an optional vector of weights to be used in the fitting
    process.  Should be \code{NULL} or a numeric vector.}

\item{family}{a description of the error distribution and link
    function to be used in the model. This can be a character string
    naming a family function, a family function or the result of a call
    to a family function.  (See \code{\link{family}} for details of
    family functions.)}

\item{offset}{this can be used to specify an \emph{a priori}
    known component to be included in the linear predictor
    during fitting.  This should be \code{NULL} or a numeric vector of
    length either one or equal to the number of cases.
    One or more \code{\link{offset}} terms can be included in the
    formula instead or as well, and if both are specified their sum is
    used.  See \code{\link{model.offset}}.}

\item{start}{starting values for the parameters in the linear predictor.}

\item{mustart}{starting values for the vector of means.}

\item{etastart}{starting values for the linear predictor.}

\item{control}{A list of control parameters including \code{maxit} for the maximum number of iterations (default is 100) and \code{epsilon} for the tolerance level (default is 1e-07).}
}
\value{
A fitted generalized linear model with Bayesian prior distributions.
}
\description{
This function serves as a wrapper around arm::bayesglm.fit,
providing additional functionality and handling various input configurations.
}
