% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_glm.R
\name{.fastglm_fit}
\alias{.fastglm_fit}
\title{A Wrapper Function for fastglm::fastglmPure}
\usage{
.fastglm_fit(
  x,
  y,
  weights = rep(1, NROW(y)),
  family = stats::gaussian(),
  offset = rep(0, NROW(y)),
  start = NULL,
  mustart = NULL,
  etastart = NULL,
  control = list(maxit = 100, epsilon = 1e-07)
)
}
\arguments{
\item{x}{input model matrix. Must be a matrix object}

\item{y}{numeric response vector of length nobs.}

\item{weights}{an optional vector of 'prior weights' to be used in the fitting process. Should be a numeric vector.}

\item{family}{a description of the error distribution and link function to be used in the model. 
For \code{fastglmPure} this can only be the result of a call to a family function. 
(See \code{\link[stats]{family}} for details of family functions.)}

\item{offset}{this can be used to specify an a priori known component to be included in the linear predictor during fitting. 
This should be a numeric vector of length equal to the number of cases}

\item{start}{starting values for the parameters in the linear predictor.}

\item{mustart}{values for the vector of means.}

\item{etastart}{starting values for the linear predictor.}

\item{control}{A list of control parameters including \code{maxit} for the maximum number of iterations (default is 100) and \code{epsilon} for the tolerance level (default is 1e-07).}
}
\value{
A fitted generalized linear model.
}
\description{
This function serves as a wrapper around fastglm::fastglmPure,
providing additional functionality and handling various input configurations.
}
