% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_similar_genes_rmd.R
\name{.load_and_filter_RmdGenesSimilarity}
\alias{.load_and_filter_RmdGenesSimilarity}
\title{Load and Filter RmdGenesSimilarity Data}
\usage{
.load_and_filter_RmdGenesSimilarity(outlierNeighborsThreshold)
}
\arguments{
\item{outlierNeighborsThreshold}{A numeric value used as the threshold for
filtering the \code{logOR} column in the data. The absolute value of
the \code{logOR} should be less than or equal to this threshold.}
}
\value{
A combined and filtered data.table object.
}
\description{
This function loads chunks of RmdGenesSimilarity data from RDA files,
combines them, and then filters the combined data based on an
outlierNeighborsThreshold value.
}
\examples{
# Example usage (assuming the function and RDA files exist)
# result_data <- .load_and_filter_RmdGenesSimilarity(0.25)

}
