% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_cadd_file.R
\name{download_cadd_file}
\alias{download_cadd_file}
\title{Download Large CADD File with Real-time Progress Indicator}
\usage{
download_cadd_file(caddScoresPath)
}
\arguments{
\item{caddScoresPath}{A character string representing the full path where the
downloaded file should be saved, including the filename.}
}
\value{
NULL. The function executes for its side effects, which include downloading
the file and saving it to \code{caddScoresPath}.
}
\description{
Downloads a substantial CADD (Combined Annotation-Dependent Depletion) file
from its official repository (https://krishna.gs.washington.edu/download/CADD/bigWig/).
The file is saved to a designated location on the user's local file system.
Utilizes the Unix \code{wget} command-line utility to manage the download process and
to provide a real-time progress update.
}
\details{
\strong{Warning}: The target file is sizeable (~11GB); ensure adequate storage space and
anticipate extended download time. Windows users might experience problems with this command. A possible solution
would be to directly download the file from https://krishna.gs.washington.edu/download/CADD/bigWig/CADD_GRCh37-v1.4.bw.
}
\examples{
\dontrun{
# Replace "your/destination/path/CADD_GRCh37-v1.4.bw" with the path
# where you want to save the file.

download_cadd_file(caddScoresPath = "your/destination/path/CADD_GRCh37-v1.4.bw")
}

}
