% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\docType{data}
\name{example_gene_annotation.csv.gz}
\alias{example_gene_annotation.csv.gz}
\title{Example Gene Annotation Data}
\format{
A compressed CSV file with columns:
\describe{
\item{Sample}{Identifier for the sample.}
\item{Gene}{Start position of the gene.}
\item{Feature}{Gene feature, for example copy number state.}
}
}
\description{
A compressed CSV file containing example gene annotation information.
}
