% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\docType{data}
\name{example_genomewide_annotation.csv.gz}
\alias{example_genomewide_annotation.csv.gz}
\title{Example Genome-wide Annotation Data}
\format{
A compressed CSV file with columns:
\describe{
\item{Sample}{Identifier for the sample.}
\item{Feature1}{Sample annotation, e.g., cancer type, treatment, etc.}
\item{Feature2}{Sample annotation, e.g., cancer type, treatment, etc.}
}
}
\description{
A compressed CSV file containing example genome-wide annotation information.
It should have at least two columns: Sample and Feature.
}
