% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\docType{data}
\name{example_mutations.csv.gz}
\alias{example_mutations.csv.gz}
\title{Example Mutations Data}
\format{
A compressed CSV file with columns:
\describe{
\item{Sample}{Identifier for the sample where the mutation was observed.}
\item{CHROM}{Chromosome where the mutation occurred.}
\item{POS}{Genomic position of the mutation.}
\item{REF}{Reference allele at the mutation site.}
\item{ALT}{Mutated (alternate) allele at the mutation site.}
}
}
\description{
A compressed CSV file containing example somatic mutations.
}
