% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_variables_from_formula.R
\name{extract_variables_from_formula}
\alias{extract_variables_from_formula}
\title{Extract Variables from a Formula String}
\usage{
extract_variables_from_formula(formula)
}
\arguments{
\item{formula}{A character string representing the formula.}
}
\value{
A character vector containing all the variables in the formula.
}
\description{
This function takes a formula string as input and returns a character vector containing all the variables in the formula.
The function does not remove any offset terms if present.
}
\examples{
\dontrun{
formula_str <- "MutationNumber ~ isTarget + CNA + Mutation + offset(log(ntAtRisk))"
extracted_vars <- extract_variables_from_formula(formula_str)
print(extracted_vars)
}

}
