% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_ranges.R
\name{filter_ranges}
\alias{filter_ranges}
\title{Filter Genomic Ranges#'}
\usage{
filter_ranges(
  grObject,
  filterRegionsPath = NULL,
  format = c("NA", "Rdata", "bed"),
  filter = c("NA", "in", "out")
)
}
\arguments{
\item{grObject}{GenomicRanges object to be filtered.}

\item{filterRegionsPath}{File path to the additional regions that need to be filtered. Default is NULL.}

\item{format}{Format of the filterRegions file, can be "Rdata", "bed", or "NA".
"NA" means no additional filtering is performed. Default is "NA".}

\item{filter}{Indicates whether to include ("in") or exclude ("out") the regions specified in filterRegionsPath.
"NA" means no additional filtering is performed. Default is "NA".}
}
\value{
A GenomicRanges object after applying the specified filters.
}
\description{
This function performs comprehensive filtering of genomic ranges, always excluding regions with low mappability
and conversion-unstable positions.
Additionally, it can optionally incorporate user-defined regions for customized filtering.
}
\examples{
\dontrun{
filterRegionsPath <- system.file("extdata",
  "HotSpotAnnotations_Trevino.RData",
  package = "mutmatch"
)
}
}
