% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_closest_neighbor.R
\name{find_closest_neighbor}
\alias{find_closest_neighbor}
\title{Identify Closest Protein-Coding Neighboring Gene}
\usage{
find_closest_neighbor(grGene, grGenes, side = c("left", "right"))
}
\arguments{
\item{grGene}{A GenomicRanges object representing the genomic coordinates of the target gene for which to find the closest neighbor.}

\item{grGenes}{A GenomicRanges object representing the genomic coordinates of the available set of genes within which to conduct the search.}

\item{side}{A character string specifying the direction of the search: 'left' for upstream or 'right' for downstream. The search is performed without taking into account the orientation of the gene on the DNA strand.}
}
\value{
Returns a vector of unique gene IDs representing the closest protein-coding neighbor(s) to the target gene. Returns NULL if no such genes are found.
}
\description{
This function searches for the nearest protein-coding gene adjacent to a specified target gene.
The search can be conducted either upstream ('left') or downstream ('right') relative to the target,
without consideration for gene directionality or orientation on the DNA strand.
}
\examples{
# Assuming `grTarget` and `grGeneSet` are defined GenomicRanges objects
\dontrun{
find_closest_neighbor(grTarget, grGeneSet, "left")
find_closest_neighbor(grTarget, grGeneSet, "right")
}
}
