% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_glm.R
\name{fit_glm}
\alias{fit_glm}
\title{Fit a Generalized Linear Model (GLM) with Various Family Options}
\usage{
fit_glm(
  formula =
    "MutationNumber ~ isTarget + CNA + isTarget:CNA + Mutation + offset(log(ntAtRisk))",
  data,
  family = c("bayes.poisson", "poisson", "negative.binomial", "bayes.negative.binomial")
)
}
\arguments{
\item{formula}{A formula object specifying the model. Default is "MutationNumber ~ isTarget + CNA + isTarget:CNA + Mutation + offset(log(ntAtRisk))".}

\item{data}{A data frame containing the variables in the model.}

\item{family}{The family for the generalized linear model ("bayes.poisson", "poisson", "negative.binomial", "bayes.negative.binomial"). Defaults to "bayes.poisson".}
}
\value{
A data table with the coefficients, p-values, and other statistics for each variable in the model.
}
\description{
This function provides a wrapper for various GLM fitting methods and returns coefficients
and other statistics of the fitted model.
}
\examples{
# Your example code here, for example:
# your_formula <- "MutationNumber ~ isTarget + CNA + isTarget:CNA+ Mutation + offset(log(ntAtRisk))"
# fit_glm(formula = your_formula, data = your_data, family = "bayes.poisson")
}
