% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cadd_model_table.R
\name{get_cadd_model_table}
\alias{get_cadd_model_table}
\title{Get CADD Regression Table}
\usage{
get_cadd_model_table(
  hgnc,
  caddScoresPath,
  mutationsPath,
  annotationGenePath,
  annotationGenomeWidePath,
  filterRegionsPath = NULL,
  format = NA,
  filter = NA
)
}
\arguments{
\item{hgnc}{The HGNC symbol of the target gene.}

\item{caddScoresPath}{Specifies the file path to a bigWig-formatted file containing CADD scores for the hg19 genome.
The default version (CADD_GRCh37-v1.4.bw) can be downloaded using the function \code{\link{download_cadd_file}}.}

\item{mutationsPath}{A path to the csv.gz file with mutations for each sample.}

\item{annotationGenePath}{A path to the file with gene-specific annotation for each sample. Default is NULL.}

\item{annotationGenomeWidePath}{A path to the file with genome-wide annotation for each sample}

\item{filterRegionsPath}{File path to the additional regions that need to be filtered. Default is NULL.}

\item{format}{Format of the filterRegions file, can be "Rdata", "bed", or "NA".
"NA" means no additional filtering is performed. Default is "NA".}

\item{filter}{Indicates whether to include ("in") or exclude ("out") the regions specified in filterRegionsPath.
"NA" means no additional filtering is performed. Default is "NA".}
}
\value{
A data.table containing categorized mutation information for the target gene and its neighboring genes.
}
\description{
Creates a mutation table for a specified target gene, categorizing mutations based on their
Combined Annotation Dependent Depletion (CADD) scores into high-CADD and low-CADD regions.
It also performs filtering based on regions.
}
\examples{
\dontrun{
# Example usage
mutation_table <- get_cadd_model_table(hgnc = "BRCA1",
caddScoresPath = "path/to/CADD_scores.bw", ...)
}

}
