% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gene_ids.R
\name{get_ensembl_from_hgnc}
\alias{get_ensembl_from_hgnc}
\title{Get Ensembl ID from HGNC Symbol}
\usage{
get_ensembl_from_hgnc(hgnc)
}
\arguments{
\item{hgnc}{Character string. The HGNC symbol for which the Ensembl ID is to be found.}
}
\value{
A string containing the Ensembl ID, or NULL if it cannot be found.
}
\description{
This function retrieves the Ensembl ID corresponding to a given HGNC symbol.
It first attempts to find this in the \code{mutmatch_topTranscripts} data frame.
If unsuccessful, it checks for approved and alias symbols.
}
