% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gene_location.R
\name{get_gene_location}
\alias{get_gene_location}
\title{Retrieve Genomic Coordinates for a Specific Gene}
\usage{
get_gene_location(ensembl_gene_id)
}
\arguments{
\item{ensembl_gene_id}{An Ensembl gene identifier}
}
\value{
A data table object with coordinates
}
\description{
This function returns the genomic coordinates associated with a specified Ensembl gene ID.
The function uses GRCh37 genome annotations to obtain this information.
It fetches the coordinates of the most highly expressed transcript of the gene,
including its exons and coding sequence (CDS), and calculates the total length
of the transcript. The function filters out genes that are not located on one of the
standard chromosomes (1-22, X, Y).
}
\examples{
\dontrun{
ensembl_gene_id <- "ENSG00000133703"
exonsCoordinates <- T
get_gene_location(ensembl_gene_id, exonsCoordinates = F)
}
}
