% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gene_neighbors.R
\name{get_gene_neighbors}
\alias{get_gene_neighbors}
\title{Identify Neighboring Genes}
\usage{
get_gene_neighbors(
  grGene,
  hgnc,
  neighborsWindow = "0.5Mb",
  outlierNeighborsThreshold = 0.2
)
}
\arguments{
\item{grGene}{The genomic coordinates of the target gene.}

\item{hgnc}{The HGNC symbol of the target gene.}

\item{neighborsWindow}{The genomic window within which to search for neighboring genes. Accepts human-readable format like "0.5 Mb".}

\item{outlierNeighborsThreshold}{The maximum absolute value of the log odds ratio of baseline mutation rate between
neighboring genes and the central gene. Genes exceeding this threshold will be filtered out when analyzing neighboring genes.}
}
\value{
A data.table of genomic coordinates for the neighboring genes, or NULL if none are found.
}
\description{
This function identifies genes neighboring a specified target gene, confined to a defined genomic window.
It filters these neighboring genes based on a similarity threshold for mutation rates relative to the
target gene. In cases where no neighbors are found within the specified window, the function selects
the closest upstream and downstream genes that meet the mutation rate similarity criteria.
}
