% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_genes_within_range.R
\name{get_genes_within_range}
\alias{get_genes_within_range}
\title{Get Genes Within a Given Range on a Chromosome}
\usage{
get_genes_within_range(chromosome, leftBorder, rightBorder)
}
\arguments{
\item{chromosome}{The name of the chromosome.}

\item{leftBorder}{The leftmost genomic coordinate.}

\item{rightBorder}{The rightmost genomic coordinate.}
}
\value{
A vector containing Ensembl IDs of genes within the specified range.
}
\description{
This function takes a chromosome name and a genomic range specified by left and right borders.
It returns the Ensembl IDs of genes that fall within this range.
}
\examples{
\dontrun{
get_genes_within_range("chr1", 1000000, 2000000)
}
}
