% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_neighbors_model_table.R
\name{get_neighbors_model_table}
\alias{get_neighbors_model_table}
\title{Get Neighbors Regression Table}
\usage{
get_neighbors_model_table(
  hgnc,
  mutationsPath,
  annotationGenePath,
  annotationGenomeWidePath,
  filterRegionsPath = NULL,
  format = NA,
  filter = NA,
  neighborsWindow = "0.5Mb",
  outlierNeighborsThreshold = 0.2
)
}
\arguments{
\item{hgnc}{The HGNC symbol of the target gene.}

\item{mutationsPath}{A path to the csv.gz file with mutations for each sample.}

\item{annotationGenePath}{A path to the file with gene-specific annotation for each sample. Default is NULL.}

\item{annotationGenomeWidePath}{A path to the file with genome-wide annotation for each sample}

\item{filterRegionsPath}{File path to the additional regions that need to be filtered. Default is NULL.}

\item{format}{Format of the filterRegions file, can be "Rdata", "bed", or "NA".
"NA" means no additional filtering is performed. Default is "NA".}

\item{filter}{Indicates whether to include ("in") or exclude ("out") the regions specified in filterRegionsPath.
"NA" means no additional filtering is performed. Default is "NA".}

\item{neighborsWindow}{The genomic window within which to search for neighboring genes. Accepts human-readable format like "0.5 Mb".}

\item{outlierNeighborsThreshold}{The maximum absolute value of the log odds ratio of baseline mutation rate between
neighboring genes and the central gene. Genes exceeding this threshold will be filtered out when analyzing neighboring genes.}
}
\value{
A data.table containing mutation information for the target gene and its neighbors.
}
\description{
This function generates a mutation table for a target gene and its neighboring genes.
It also performs filtering based on regions and annotations.
}
\examples{
\dontrun{
# Example usage
mutation_table <- get_neighbors_model_table(...)
}
}
