% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_selection_estimates_cadd.R
\name{get_selection_estimates_cadd}
\alias{get_selection_estimates_cadd}
\title{Estimate Selection Pressures on a Target Gene Using Low-CADD Gene Regions as a Mutational Rate Baseline}
\usage{
get_selection_estimates_cadd(
  hgnc,
  caddScoresPath,
  mutationsPath,
  annotationGenePath,
  annotationGenomeWidePath,
  filterRegionsPath = NULL,
  format = NA,
  filter = NA,
  formula =
    "MutationNumber ~ isTarget + CNA + isTarget:CNA + Mutation + offset(log(ntAtRisk))",
  family = "bayes.poisson",
  simtimes = 50
)
}
\arguments{
\item{hgnc}{The HGNC symbol of the target gene.}

\item{caddScoresPath}{Specifies the file path to a bigWig-formatted file containing CADD scores for the hg19 genome.
The default version (CADD_GRCh37-v1.4.bw) can be downloaded using the function \code{\link{download_cadd_file}}.}

\item{mutationsPath}{A path to the csv.gz file with mutations for each sample.}

\item{annotationGenePath}{A path to the file with gene-specific annotation for each sample. Default is NULL.}

\item{annotationGenomeWidePath}{A path to the file with genome-wide annotation for each sample}

\item{filterRegionsPath}{File path to the additional regions that need to be filtered. Default is NULL.}

\item{format}{Format of the filterRegions file, can be "Rdata", "bed", or "NA".
"NA" means no additional filtering is performed. Default is "NA".}

\item{filter}{Indicates whether to include ("in") or exclude ("out") the regions specified in filterRegionsPath.
"NA" means no additional filtering is performed. Default is "NA".}

\item{formula}{A formula object specifying the model. Default is "MutationNumber ~ isTarget + CNA + isTarget:CNA + Mutation + offset(log(ntAtRisk))".}

\item{family}{The family for the generalized linear model ("bayes.poisson", "poisson", "negative.binomial", "bayes.negative.binomial"). Defaults to "bayes.poisson".}

\item{simtimes}{The number of times to simulate the data for correction of selection
estimates. Defaults to 50.}
}
\value{
A data.table containing the debiased selection estimates, including estimated coefficients and their uncertainties.
}
\description{
Performs a comprehensive two-stage analysis to quantify the selection forces acting upon a specified target gene.
In the first stage, the function separates gene regions with low-CADD and high-CADD scores.
In the second stage, a Bayesian Poisson regression model is employed to derive selection estimates for the target gene.
These estimates are subsequently refined through permutation-based debiasing.
}
\examples{
\dontrun{
# Replace "your/destination/path/CADD_GRCh37-v1.4.bw" with the path where
you want to save the file.

caddScoresPath = "your/destination/path/CADD_GRCh37-v1.4.bw"
download_cadd_file(caddScoresPath = caddScoresPath)

selection_estimates <- get_selection_estimates_cadd(
  hgnc = "KRAS",
  caddScoresPath = caddScoresPath,
  mutationsPath = system.file("extdata",
    "example_mutations.csv.gz",
    package = "mutmatch"
  ),
  annotationGenePath = system.file("extdata",
  "example_gene_annotation.csv.gz",
    package = "mutmatch"
  ),
  annotationGenomeWidePath = system.file("extdata",
    "example_genomewide_annotation.csv",
    package = "mutmatch"
  )
)
}
}
