% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_similar_genes_rmd.R
\name{get_similar_genes_rmd}
\alias{get_similar_genes_rmd}
\title{Get Similar Genes Based on Rmd (Regional Mutation Density)}
\usage{
get_similar_genes_rmd(hgnc, outlierNeighborsThreshold)
}
\arguments{
\item{hgnc}{The HGNC symbol of the target gene.}

\item{outlierNeighborsThreshold}{The maximum absolute value of the log odds ratio of baseline mutation rate between
neighboring genes and the central gene. Genes exceeding this threshold will be filtered out when analyzing neighboring genes.}
}
\value{
character vector of unique GeneIDs that are similar to the specified gene.
}
\description{
This function retrieves genes that are similar to a specified gene
(\code{hgnc}) based on a provided threshold (\code{outlierNeighborsThreshold}).
The function filters the data to return only the GeneIDs that are similar
to the specified gene based on the \code{logOR} value.
}
\examples{
# Example usage (assuming the function and relevant data exist)
similar_genes <- get_similar_genes_rmd("KRAS", 0.2)
}
