% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_mutations.R
\name{load_mutations}
\alias{load_mutations}
\title{Load and Preprocess Mutation Data}
\usage{
load_mutations(mutationsPath, chr = NULL)
}
\arguments{
\item{mutationsPath}{A path to the csv.gz file with mutations for each sample.}

\item{chr}{Optional. The chromosome for which mutations should be specifically loaded.
If not specified, mutations for all chromosomes are loaded.}
}
\value{
A GRanges object containing filtered and preprocessed mutations data.
}
\description{
This function reads in mutation data from a specified file path and performs several preprocessing steps.
It offers the option to focus only on mutations for a specific chromosome.
The function also converts the mutation table to a GRanges object, retains important columns, and standardizes chromosome naming.
Finally, it filters out any SNVs that don't meet specific length criteria for the reference and mutated alleles.
}
\examples{
\dontrun{
grSNVs <- load_mutations(mutationsPath = "example_mutations.csv.gz")
grSNVs_chr1 <- load_mutations(
  mutationsPath = "example_mutations.csv.gz",
  chr = "1"
)
grSNVs_chr1 <- load_mutations(mutationsPath = system.file("extdata",
  "example_mutations.csv.gz",
  package = "mutmatch"
), chr = "1")
}
}
