% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_ranges_from_table.R
\name{make_ranges_from_table}
\alias{make_ranges_from_table}
\title{Create Genomic Ranges Object from Genes Data Table}
\usage{
make_ranges_from_table(rangesData, exonsCoordinates = TRUE)
}
\arguments{
\item{rangesData}{A data frame containing columns 'chromosome_name', 'genomic_coding_start', 'genomic_coding_end',
'strand', 'length', 'start_position', and 'end_position'.}

\item{exonsCoordinates}{A logical flag specifying if exon coordinates should be used. Default is TRUE.}
}
\value{
A GenomicRanges object containing the gene information.
}
\description{
This function takes a data table containing gene information, such as genomic start, genomic end, and chromosome,
and returns a GenomicRanges object.
}
\examples{
\dontrun{
genesData <- data.frame(
  chromosome_name = c("1", "1"),
  genomic_coding_start = c(1000, 2000),
  genomic_coding_end = c(1500, 2500),
  strand = c("+", "-"),
  length = c(500, 500),
  start_position = c(1000, 2000),
  end_position = c(1500, 2500)
)
grGenesData <- make_ranges_from_table(genesData)
}

}
