% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_sql_gene_annotation.R
\name{make_sql_gene_annotation}
\alias{make_sql_gene_annotation}
\title{Make SQL Gene Annotation Database}
\usage{
make_sql_gene_annotation(annotationGenePath, annotationGenePathSQL)
}
\arguments{
\item{annotationGenePath}{A character string specifying the path to the file containing gene annotation data.
The file at this path will be read into R.}

\item{annotationGenePathSQL}{A character string specifying the path where the SQL database will be stored.
This path must end with a '.sqlite' or '.db' extension to indicate that it is an SQLite database file.}
}
\description{
This function reads gene annotation data from a specified file and writes it to an SQLite database.
}
\examples{
\dontrun{
make_sql_gene_annotation(
  annotationGenePath = system.file("extdata", "example_gene_annotation.csv.gz", package = "mutmatch"),
  annotationGenePathSQL = "/path/to/example_gene_annotation.sqlite"
)
}
}
