% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutmatch-package.R
\docType{package}
\name{mutmatch-package}
\alias{mutmatch}
\alias{mutmatch-package}
\title{mutmatch: Estimating Somatic Selection Using Neighbors Baseline or Low-Impact Mutations}
\description{
Provides a suite of functions to estimate somatic selection for genes by utilizing the baseline mutation rates of neighboring genes or low-impact mutations within the same gene. Incorporates CADD score for low-impact mutation analysis.
}
\author{
\strong{Maintainer}: Elizaveta Besedina \email{bio.besedina@gmail.com} (\href{https://orcid.org/0000-0002-8575-6688}{ORCID})

}
\keyword{internal}
