% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_process_pvalues.R
\name{post_process_pvalues}
\alias{post_process_pvalues}
\title{Correct P-Values for Selection Estimates Across Multiple Genes}
\usage{
post_process_pvalues(selection_estimates)
}
\arguments{
\item{selection_estimates}{A data frame of selection estimates that includes columns for real and simulated p-values.}
}
\value{
A data frame with corrected p-values for the specified coefficients, enhancing the reliability of selection analyses.
}
\description{
This function refines p-value estimates for selection analyses by correcting biases
across a comprehensive range of genes. It applies the \code{correct_pvalues} function to adjust p-values for
specific coefficients (i.e., \code{isTarget1}, \code{isTarget1:CNA-1}, and \code{isTarget1:CNA1}) in the selection estimates dataset.
For enhanced accuracy, it's recommended to apply this function to data encompassing a broad spectrum of genes.
Alternatively, increase the simulation count (\code{simtimes} parameter) in functions like \code{get_selection_estimates_neighbors}
or \code{get_selection_estimates_cadd} to improve resolution when extensive gene data is unavailable.
}
\examples{
\dontrun{
selection_estimates <- get_selection_estimates_neighbors(
hgnc = "KRAS",
mutationsPath = system.file("extdata", "example_mutations.csv.gz", package = "mutmatch"),
annotationGenePath = system.file("extdata", "example_gene_annotation.csv.gz", package = "mutmatch"),
annotationGenomeWidePath = system.file("extdata", "example_genomewide_annotation.csv", package = "mutmatch"),
neighborsWindow = "0.5Mb",
outlierNeighborsThreshold = 0.2,
simtimes = 500
)
 selection_estimates <- post_process_pvalues(selection_estimates)
 }
}
