% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_mismatched_samples.R
\name{remove_mismatched_samples}
\alias{remove_mismatched_samples}
\title{Remove Mismatched Samples}
\usage{
remove_mismatched_samples(genes, annotationGenePath, referenceAnnotation)
}
\arguments{
\item{genes}{A vector of HGNC gene symbols to be used for filtering samples.}

\item{annotationGenePath}{A path to the file with gene-specific annotation for each sample.
Can be 'sqlite' or 'db' if the file is produced with make_sql_gene_annotation function (recommended for large files). Default is NULL.}

\item{referenceAnnotation}{Reference annotation from which mismatched samples have to be removed.}
}
\value{
A modified version of the \code{referenceAnnotation} data.table or data.frame, containing only samples
that match the gene-specific annotations.
}
\description{
This function removes samples from a reference annotation table whose gene-specific annotations
do not match those in the provided list of genes.
}
\examples{
# Usage:
# filtered_reference = remove_mismatched_samples(genes, annotationGenePath, referenceAnnotation)
}
