% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shuffle_mutation_counts.R
\name{shuffle_mutation_counts}
\alias{shuffle_mutation_counts}
\title{Shuffle Mutation Counts for a Single Cohort}
\usage{
shuffle_mutation_counts(mutation_table, ntimes = 50)
}
\arguments{
\item{mutation_table}{A data.table containing mutation information. It should contain only one category (cohort)
except for the 'isTarget' and 'Mutation' variables.}

\item{ntimes}{The number of times the shuffling is to be done.}
}
\value{
A data.table containing the permuted mutation counts.
}
\description{
This function takes a mutation table for a single cohort and returns a table with permuted mutation counts.
The shuffling is done based on the marginal sums of mutations and ntAtRisk for each mutation type.
The returned table will have a new column named 'permutation_id' to identify each permutation.
}
\examples{
\dontrun{
shuffled_data <- shuffle_mutation_counts(mutation_table, ntimes = 100)
}
}
