% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_selection_model.R
\name{fit_selection_model}
\alias{fit_selection_model}
\title{Fit Selection Model and Debias Its Estimates}
\usage{
fit_selection_model(
  mutation_table,
  formula =
    "MutationNumber ~ isTarget + CNA + isTarget:CNA + Mutation + offset(log(ntAtRisk))",
  family = "bayes.poisson",
  simtimes = 50
)
}
\arguments{
\item{mutation_table}{A data.table containing the mutation data. The table should
include mandatory columns 'MutationNumber' and 'isTarget', as well as any additional
covariates that may be specified in the model formula. If any covariates are present
in the data but not specified in the formula, the data will be subsetted according to
these covariates to fit individual models for each subset.}

\item{formula}{A formula object specifying the model. Default is "MutationNumber ~ isTarget + CNA + isTarget:CNA + Mutation + offset(log(ntAtRisk))".}

\item{family}{The family for the generalized linear model ("bayes.poisson", "poisson", "negative.binomial", "bayes.negative.binomial"). Defaults to "bayes.poisson".}

\item{simtimes}{The number of times for bias correction. Defaults to 50.}
}
\value{
A data.table containing corrected estimates.
}
\description{
This function fits a selection model to provided mutation data based on a user-defined formula.
The process involves an initial simulation of the data under a neutral selection model.
Following this, the chosen model is applied to both the real and simulated datasets.
The output estimates are then corrected using the simulated data as a reference.
Additionally, the function aggregates and provides the total number of mutations
for both tested and controlled regions across distinct cohorts.
}
