% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_selection_estimates_neighbors.R
\name{get_selection_estimates_neighbors}
\alias{get_selection_estimates_neighbors}
\title{Estimate Selection Pressures on a Target Gene Using Neighboring Genes as a Mutational Rate Baseline}
\usage{
get_selection_estimates_neighbors(
  hgnc,
  mutationsPath,
  annotationGenePath,
  annotationGenomeWidePath,
  filterRegionsPath = NULL,
  format = NA,
  filter = NA,
  neighborsWindow = "0.5Mb",
  outlierNeighborsThreshold = 0.2,
  formula =
    "MutationNumber ~ isTarget + CNA + isTarget:CNA + Mutation + offset(log(ntAtRisk))",
  family = "bayes.poisson",
  simtimes = 50
)
}
\arguments{
\item{hgnc}{The HGNC symbol of the target gene.}

\item{mutationsPath}{A path to the csv.gz file with mutations for each sample.}

\item{annotationGenePath}{A path to the file with gene-specific annotation for each sample. Default is NULL.}

\item{annotationGenomeWidePath}{A path to the file with genome-wide annotation for each sample}

\item{filterRegionsPath}{File path to the additional regions that need to be filtered. Default is NULL.}

\item{format}{Format of the filterRegions file, can be "Rdata", "bed", or "NA".
"NA" means no additional filtering is performed. Default is "NA".}

\item{filter}{Indicates whether to include ("in") or exclude ("out") the regions specified in filterRegionsPath.
"NA" means no additional filtering is performed. Default is "NA".}

\item{neighborsWindow}{The genomic window within which to search for neighboring genes. Accepts human-readable format like "0.5 Mb".}

\item{outlierNeighborsThreshold}{The maximum absolute value of the log odds ratio of baseline mutation rate between
neighboring genes and the central gene. Genes exceeding this threshold will be filtered out when analyzing neighboring genes.}

\item{formula}{A formula object specifying the model. Default is "MutationNumber ~ isTarget + CNA + isTarget:CNA + Mutation + offset(log(ntAtRisk))".}

\item{family}{The family for the generalized linear model ("bayes.poisson", "poisson", "negative.binomial", "bayes.negative.binomial"). Defaults to "bayes.poisson".}

\item{simtimes}{The number of times for bias correction. Defaults to 50.}
}
\value{
A data.table containing the debiased selection estimates, including estimated coefficients and their uncertainties.
}
\description{
Performs a comprehensive two-stage analysis to quantify the selection forces acting upon a specified target gene.
In the first stage, the function identifies and filters neighboring genes within a user-defined genomic window based on their
mutation rate similarity.
In the second stage, a Bayesian Poisson regression model is employed to derive selection estimates for the target gene.
These estimates are subsequently refined through permutation-based debiasing.
}
\examples{
\dontrun{
selection_estimates <- get_selection_estimates_neighbors(
  hgnc = "KRAS",
  mutationsPath = system.file("extdata",
    "example_mutations.csv.gz",
    package = "mutmatch"
  ),
  annotationGenePath = system.file("extdata",
    "example_gene_annotation.csv.gz",
    package = "mutmatch"
  ),
  annotationGenomeWidePath = system.file("extdata",
    "example_genomewide_annotation.csv",
    package = "mutmatch"
  )
)
}
}
